Avoid more torch imports
